Moved mongo specific code out of app.py

+++ data.py
@@ -30,6 +30,9 @@ def get_books(params):
 
 
 def get_book(obj_id):
+    if not ObjectId.is_valid(obj_id):
+        return 'Invalid'
+
     result = client[database][collection].find_one({"_id": ObjectId(obj_id)})
     if result is not None:
         result['_id'] = str(result['_id'])
@@ -37,17 +40,26 @@ def get_book(obj_id):
     return result
 
 
-def post_book(body):
-    if "_id" in body:
-        body['_id'] = ObjectId(body['_id'])
+def create_book(body):
+    if '_id' in body:
+        if not ObjectId.is_valid(body['_id']):
+            return None
+        else:
+            body['_id'] = ObjectId(body['_id'])
 
     return client[database][collection].insert_one(body)
 
 
-def put_book(obj_id, body):
+def update_book(obj_id, body):
+    if not ObjectId.is_valid(obj_id):
+        return None
+
     body.pop('_id', None)
     return client[database][collection].update_one({"_id": ObjectId(obj_id)}, {'$set': body}, upsert=True)
 
 
 def delete_book(obj_id):
-    client[database][collection].delete_one({"_id": ObjectId(obj_id)})
+    if not ObjectId.is_valid(obj_id):
+        return None
+
+    return client[database][collection].delete_one({"_id": ObjectId(obj_id)})
